RENAME #60: global 폴더에 image 폴더 생성 및 이미지 처리 관련 클래스 image 폴더로 위치 변경
